feat(424): AI offering-sections extraction on priced prospectuses (424B1/424B4)
Priced prospectuses now record the FINAL deal under extractor id '424':
offering terms (equity -> offering_terms, SPAC units -> spac_unit_terms),
the exact ticker series, underwriters rolled up to families, and
use-of-proceeds line items — alongside the S-1's registered/anticipated
terms, so the registration -> pricing evolution is queryable per accession.
Shelf takedowns and supplements (424A, 424B2/B3/B5/B7) stay
deterministic-only (XBRL pass + issuer observation).

To share the pipeline rather than duplicate it, the per-section
dead-letter ceremony moves to s1/sectionRunner.ts (makeRunSection) and the
three offering sections to s1/offeringSections.ts (runOfferingSections),
both now used by processFormS1 and processForm424. SPAC classification for
the 424 branch comes from the SGML-header SIC (6770), same as the S-1.
HTML-converter throws dead-letter the offering sections as PARSE_ERROR,
mirroring the S-1 containment.
