+++ src/sec/forms/registration-statements/Form_S_1.storage.ts
@@ -20,15 +20,6 @@ import { CanonicalSponsorFamilyAliasRepo } from "../../../storage/canonical/Cano
 import { SponsorFamilyResolver } from "../../../resolver/SponsorFamilyResolver";
 import { SponsorFamilyMembershipRepo } from "../../../storage/canonical/SponsorFamilyMembershipRepo";
 import { SpacSponsorLinkRepo } from "../../../storage/canonical/SpacSponsorLinkRepo";
-import { OfferingTermsRepo } from "../../../storage/offering/OfferingTermsRepo";
-import { SpacUnitTermsRepo } from "../../../storage/offering/SpacUnitTermsRepo";
-import { IssuerTickerRepo } from "../../../storage/offering/IssuerTickerRepo";
-import { CanonicalUnderwriterFamilyRepo } from "../../../storage/canonical/CanonicalUnderwriterFamilyRepo";
-import { CanonicalUnderwriterFamilyAliasRepo } from "../../../storage/canonical/CanonicalUnderwriterFamilyAliasRepo";
-import { UnderwriterFamilyResolver } from "../../../resolver/UnderwriterFamilyResolver";
-import { UnderwriterFamilyMembershipRepo } from "../../../storage/canonical/UnderwriterFamilyMembershipRepo";
-import { UnderwriterLinkRepo } from "../../../storage/canonical/UnderwriterLinkRepo";
-import { UseOfProceedsRepo } from "../../../storage/use-of-proceeds/UseOfProceedsRepo";
 import type { FormS1Parsed } from "./Form_S_1";
 import { parseEdgarHtml } from "../../html/parseEdgarHtml";
 import { DocumentTreeSegmenter } from "./s1/DocumentTreeSegmenter";
@@ -37,12 +28,11 @@ import { spanAppearsIn } from "./s1/verifySourceSpan";
 import {
   extractBeneficialOwnership,
   extractManagement,
-  extractOfferingTerms,
   extractRelatedParty,
   extractSpacSponsors,
-  extractUnderwriters,
-  extractUseOfProceeds,
 } from "./s1/sectionExtractors";
+import { makeRunSection } from "./s1/sectionRunner";
+import { runOfferingSections } from "./s1/offeringSections";
 import { getS1Model } from "./s1/s1Model";
 import { splitPersonName } from "./s1/splitName";
 import { extractAndStoreXbrl } from "./s1/xbrlEnrichment";
@@ -52,20 +42,6 @@ const EXTRACTOR_ID = "S-1";
 // appear verbatim (after light normalization) in the section text before a
 // canonical sponsor row is persisted.
 const DEFAULT_EXTRACTOR_VERSION = "1.1.0";
-const RAW_CONFIDENCE_FLOOR = Number(process.env.SEC_S1_CONFIDENCE_FLOOR ?? "0");
-// A non-numeric SEC_S1_CONFIDENCE_FLOOR would be NaN, and `confidence >= NaN` is
-// always false — silently dropping every row. Fall back to 0 (no floor).
-const CONFIDENCE_FLOOR = Number.isFinite(RAW_CONFIDENCE_FLOOR) ? RAW_CONFIDENCE_FLOOR : 0;
-
-/**
- * Share/unit counts are emitted by the model as plain numbers but stored in
- * integer-typed columns. Round a finite value to the nearest integer (a stray
- * decimal would otherwise be rejected on write and dead-letter the whole
- * section); pass through null.
- */
-function toIntCount(n: number | null | undefined): number | null {
-  return n == null || !Number.isFinite(n) ? null : Math.round(n);
-}
 
 export interface ProcessFormS1Args {
   readonly cik: number;
@@ -126,25 +102,9 @@ export async function processFormS1(args: ProcessFormS1Args): Promise<void> {
   const membershipRepo = new SponsorFamilyMembershipRepo();
   const linkRepo = new SpacSponsorLinkRepo();
 
-  const offeringTermsRepo = new OfferingTermsRepo();
-  const spacUnitTermsRepo = new SpacUnitTermsRepo();
-  const issuerTickerRepo = new IssuerTickerRepo();
-
-  const underwriterFamilyResolver = new UnderwriterFamilyResolver({
-    canonicalUnderwriterFamilyRepo: new CanonicalUnderwriterFamilyRepo(),
-    canonicalUnderwriterFamilyAliasRepo: new CanonicalUnderwriterFamilyAliasRepo(),
-    activeResolverVersion: activeUnderwriterFamilyVersion,
-  });
-  const underwriterMembershipRepo = new UnderwriterFamilyMembershipRepo();
-  const underwriterLinkRepo = new UnderwriterLinkRepo();
-  const useOfProceedsRepo = new UseOfProceedsRepo();
-
   await ownershipRepo.clear(accession_number);
   await relatedRepo.clear(accession_number);
   await linkRepo.clear(accession_number);
-  await issuerTickerRepo.clear(accession_number);
-  await underwriterLinkRepo.clear(accession_number);
-  await useOfProceedsRepo.clear(accession_number);
 
   const base = { accession_number, extractor_id: EXTRACTOR_ID, extractor_version };
   let idx = 0;
@@ -223,118 +183,12 @@ export async function processFormS1(args: ProcessFormS1Args): Promise<void> {
   const recordOk = (section: S1SectionName) =>
     deadLetters.markResolved(EXTRACTOR_ID, accession_number, section);
 
-  /**
-   * Shared per-section ceremony: resolve text, dead-letter when absent, run the
-   * extractor, apply the confidence floor, persist surviving rows, and emit the
-   * resolved / empty / low-confidence / invalid-output dead letters. All seven
-   * S-1 sections funnel through here so the policy lives in exactly one place.
-   *
-   * `section_name` doubles as the dead-letter `section_name`; it is the
-   * `S1SectionName` for the entity sections and a literal string for the
-   * derived offering / underwriter / proceeds / sponsor sections.
-   */
-  async function runSection<TRow extends { confidence: number }>(sargs: {
-    sectionName: string;
-    text: string | undefined;
-    skip?: boolean;
-    notFoundDetail?: string | null;
-    emptyDetail: string;
-    lowConfidenceDetail: string;
-    // When set, a persist that writes 0 of N rows (e.g. all underwriter/sponsor
-    // rows had blank names) dead-letters MODEL_INVALID_OUTPUT. Omit for sections
-    // whose persist always writes every confident row, so they always markResolved.
-    invalidWriteDetail?: string;
-    // Optional row-level verification applied AFTER the confidence floor. When
-    // every confident row is dropped, the section dead-letters as
-    // UNVERIFIED_SOURCE_SPAN (using `unverifiedAllDetail`); when some are
-    // dropped, the surviving rows persist normally AND a "<sectionName>-partial"
-    // dead-letter is recorded for triage (using `unverifiedPartialDetail`).
-    // Detail strings may use `$N` (dropped count) and `$T` (confident total).
-    // `NoInfer<TRow>` keeps TRow inferred solely from `extract` — without it,
-    // contextual typing of the verifyRow callback's parameter would pin TRow
-    // to the constraint and break the persist callback's row typing.
-    verifyRow?: (text: string, row: NoInfer<TRow>) => boolean;
-    unverifiedAllDetail?: string;
-    unverifiedPartialDetail?: string;
-    extract: (text: string) => Promise<TRow[]>;
-    persist: (rows: TRow[]) => Promise<number>;
-  }): Promise<void> {
-    if (sargs.skip) return;
-
-    const record = (reason: string, detail: string | null) =>
-      deadLetters.record({
-        extractor_id: EXTRACTOR_ID,
-        accession_number,
-        section_name: sargs.sectionName,
-        reason_code: reason,
-        detail,
-        failed_extractor_version: extractor_version,
-        source_run_id: null,
-      });
-
-    if (sargs.text === undefined || sargs.text.trim() === "") {
-      await record("SECTION_NOT_FOUND", sargs.notFoundDetail ?? null);
-      return;
-    }
-
-    try {
-      const raw = await sargs.extract(sargs.text);
-      const confident = raw.filter((r) => r.confidence >= CONFIDENCE_FLOOR);
-      const text = sargs.text;
-      const verifyRow = sargs.verifyRow;
-      let rows: TRow[];
-      let droppedUnverified = 0;
-      if (verifyRow !== undefined && confident.length > 0) {
-        rows = confident.filter((r) => verifyRow(text, r));
-        droppedUnverified = confident.length - rows.length;
-      } else {
-        rows = confident;
-      }
-      if (rows.length === 0) {
-        const allDroppedUnverified =
-          droppedUnverified > 0 && droppedUnverified === confident.length;
-        const reason = allDroppedUnverified
-          ? "UNVERIFIED_SOURCE_SPAN"
-          : raw.length === 0
-            ? "MODEL_EMPTY"
-            : "LOW_CONFIDENCE_ALL";
-        const detail = allDroppedUnverified
-          ? (sargs.unverifiedAllDetail ?? sargs.lowConfidenceDetail).replace(
-              /\$T/g,
-              String(confident.length)
-            )
-          : raw.length === 0
-            ? sargs.emptyDetail
-            : sargs.lowConfidenceDetail;
-        await record(reason, detail);
-        return;
-      }
-      const wrote = await sargs.persist(rows);
-      if (sargs.invalidWriteDetail !== undefined && wrote === 0) {
-        await record("MODEL_INVALID_OUTPUT", sargs.invalidWriteDetail);
-      } else {
-        await deadLetters.markResolved(EXTRACTOR_ID, accession_number, sargs.sectionName);
-      }
-      if (droppedUnverified > 0 && sargs.unverifiedPartialDetail !== undefined) {
-        await deadLetters.record({
-          extractor_id: EXTRACTOR_ID,
-          accession_number,
-          section_name: `${sargs.sectionName}-partial`,
-          reason_code: "UNVERIFIED_SOURCE_SPAN",
-          detail: sargs.unverifiedPartialDetail
-            .replace(/\$N/g, String(droppedUnverified))
-            .replace(/\$T/g, String(confident.length)),
-          failed_extractor_version: extractor_version,
-          source_run_id: null,
-        });
-      }
-    } catch (e) {
-      await record(
-        "MODEL_INVALID_OUTPUT",
-        (e instanceof Error ? e.message : String(e)).slice(0, 1024)
-      );
-    }
-  }
+  const runSection = makeRunSection({
+    deadLetters,
+    extractor_id: EXTRACTOR_ID,
+    extractor_version,
+    accession_number,
+  });
 
   // The entity sections feed a SECTION_NOT_FOUND with a `null` detail when the
   // text is undefined. `runSection` also treats a blank string as not-found,
@@ -503,170 +357,21 @@ export async function processFormS1(args: ProcessFormS1Args): Promise<void> {
     },
   });
 
-  // --- Offering terms (read from The Offering + Underwriting) ---
-  // The extractor returns a single object; adapt it onto runSection by treating
-  // a null result as an empty array and wrapping a present result as `[terms]`.
-  const offeringText = [byName.get(S1_SECTIONS.THE_OFFERING), byName.get(S1_SECTIONS.UNDERWRITING)]
-    .filter((t): t is string => typeof t === "string")
-    .join("\n\n");
-  await runSection({
-    sectionName: "offering-terms",
-    text: offeringText,
-    notFoundDetail: "no The Offering / Underwriting section text",
-    emptyDetail: "no offering terms returned",
-    lowConfidenceDetail: "below confidence floor",
-    extract: async (text) => {
-      const terms = await extractOfferingTerms(text, model);
-      return terms === null ? [] : [terms];
-    },
-    persist: async (rows) => {
-      const terms = rows[0];
-      const now = new Date().toISOString();
-      if (isSpac) {
-        await spacUnitTermsRepo.save({
-          extractor_id: EXTRACTOR_ID,
-          accession_number,
-          cik,
-          units_offered: toIntCount(terms.units_offered),
-          price_per_unit: terms.price_per_unit,
-          unit_composition: terms.unit_composition,
-          warrant_fraction_per_unit: terms.warrant_fraction_per_unit,
-          right_fraction_per_unit: terms.right_fraction_per_unit,
-          trust_per_unit: terms.trust_per_unit,
-          over_allotment_units: toIntCount(terms.over_allotment_units),
-          exchange: terms.exchange,
-          ticker: terms.tickers.find((t) => t.is_primary)?.ticker ?? null,
-          gross_proceeds: terms.gross_proceeds,
-          net_proceeds: terms.net_proceeds,
-          confidence: terms.confidence,
-          source_span: terms.source_span,
-          created_at: now,
-        });
-      } else {
-        await offeringTermsRepo.save({
-          extractor_id: EXTRACTOR_ID,
-          accession_number,
-          cik,
-          security_type: terms.security_type,
-          shares_offered: toIntCount(terms.shares_offered),
-          price: terms.price,
-          price_low: terms.price_low,
-          price_high: terms.price_high,
-          gross_proceeds: terms.gross_proceeds,
-          net_proceeds: terms.net_proceeds,
-          over_allotment_shares: toIntCount(terms.over_allotment_shares),
-          exchange: terms.exchange,
-          ticker: terms.tickers.find((t) => t.is_primary)?.ticker ?? null,
-          par_value: terms.par_value,
-          confidence: terms.confidence,
-          source_span: terms.source_span,
-          created_at: now,
-        });
-      }
-      for (const t of terms.tickers) {
-        const ticker = t.ticker?.trim() ?? "";
-        if (ticker === "") continue;
-        await issuerTickerRepo.save({
-          extractor_id: EXTRACTOR_ID,
-          accession_number,
-          exchange: (t.exchange ?? terms.exchange ?? "").trim(),
-          ticker,
-          cik,
-          filing_date: args.filing_date,
-          security_type: t.security_type,
-          is_primary: t.is_primary,
-          confidence: terms.confidence,
-          source_span: terms.source_span,
-          created_at: now,
-        });
-      }
-      return 1;
-    },
-  });
-
-  // --- Underwriters (Underwriting section; all filings) ---
-  await runSection({
-    sectionName: "underwriters",
-    text: byName.get(S1_SECTIONS.UNDERWRITING),
-    emptyDetail: "no underwriters returned",
-    lowConfidenceDetail: "all rows below confidence floor",
-    invalidWriteDetail: "no underwriter rows had usable legal and common names",
-    extract: (text) => extractUnderwriters(text, model),
-    persist: async (rows) => {
-      let wrote = 0;
-      for (const r of rows) {
-        const legalName = r.legal_name?.trim() ?? "";
-        const commonName = r.common_name?.trim() ?? "";
-        if (legalName === "" || commonName === "") continue;
-        const observation_index = idx++;
-        const { observation_id, canonical_company_id } = await observer.observeCompany({
-          ...base,
-          observation_index,
-          name: legalName,
-          source_context: JSON.stringify({ relation: "s1:underwriter" }),
-        });
-        await provenance.save({
-          kind: "company",
-          observation_id,
-          confidence: r.confidence,
-          source_span: r.source_span,
-          section_name: "underwriters",
-          model_id,
-          prompt_version: extractor_version,
-          extra: null,
-        });
-        const underwriter_family_id = await underwriterFamilyResolver.resolve(commonName);
-        await underwriterMembershipRepo.record({
-          resolver_version: activeUnderwriterFamilyVersion,
-          canonical_company_id,
-          canonical_underwriter_family_id: underwriter_family_id,
-          seen_at: new Date().toISOString(),
-        });
-        await underwriterLinkRepo.save({
-          accession_number,
-          extractor_id: EXTRACTOR_ID,
-          observation_index,
-          issuer_cik: cik,
-          underwriter_canonical_company_id: canonical_company_id,
-          underwriter_family_id,
-          role_detail: r.role,
-          shares_allocated: toIntCount(r.shares_allocated),
-          over_allotment_shares: toIntCount(r.over_allotment_shares),
-          resolver_version: activeUnderwriterFamilyVersion,
-        });
-        wrote++;
-      }
-      return wrote;
-    },
-  });
-
-  // --- Use of proceeds ---
-  await runSection({
-    sectionName: "use-of-proceeds",
-    text: byName.get(S1_SECTIONS.USE_OF_PROCEEDS),
-    emptyDetail: "no line items returned",
-    lowConfidenceDetail: "all rows below confidence floor",
-    extract: (text) => extractUseOfProceeds(text, model),
-    persist: async (rows) => {
-      const now = new Date().toISOString();
-      let lineIndex = 0;
-      for (const r of rows) {
-        await useOfProceedsRepo.save({
-          extractor_id: EXTRACTOR_ID,
-          accession_number,
-          line_index: lineIndex++,
-          cik,
-          purpose: r.purpose,
-          amount: r.amount,
-          percent: r.percent,
-          note: r.note,
-          confidence: r.confidence,
-          source_span: r.source_span,
-          created_at: now,
-        });
-      }
-      return rows.length;
-    },
+  await runOfferingSections({
+    runSection,
+    observer,
+    provenance,
+    nextIndex: () => idx++,
+    accession_number,
+    extractor_id: EXTRACTOR_ID,
+    extractor_version,
+    cik,
+    filing_date: args.filing_date,
+    isSpac,
+    model,
+    model_id,
+    activeUnderwriterFamilyVersion,
+    byName,
   });
 
   // --- SPAC sponsors (gated on deterministic classification) ---
